wmppp.app: Display speed in K when too high.
When the download speed is too high, wmppp can't display it because it has
only 5 digit. The most significant digits are hidden !

This small patch display the speed in K when it is too high.

Based on the patch by jguiton <jguiton@free.fr>.
See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=328699.